requirements.txt
# Django Core
Django>=4.2.0,<5.0
django-environ>=0.11.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Image Processing
Pillow>=10.0.0

# Static Files (Production)
whitenoise>=6.5.0

# WSGI Server (Production)
gunicorn>=21.2.0

# AWS S3 Storage (Optional - for production media files)
django-storages>=1.14.0
boto3>=1.28.0

# Phone Number Validation
phonenumbers>=8.13.0
django-phonenumber-field>=7.0.0

# Forms & Admin
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Internationalization
django-modeltranslation>=0.18.0

# Security
django-cors-headers>=4.3.0

# Development Tools
ipython>=8.15.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0

# Date/Time Utilities
python-dateutil>=2.8.0

